refactor: modularize useWrapButton
